Added ability to detect multiple faces
