+++ src/App.js
@@ -39,21 +39,22 @@ class App extends Component {
   }
 
   calculateFaceLocation = (data) => {
-    const clarifaiface = data.outputs[0].data.regions[0].region_info.bounding_box;
-    const image = document.getElementById('inputimage');
-    const width = Number(image.width);
-    const height = Number(image.height);
-    console.log(width,height);
-    return{
-      leftCol: clarifaiface.left_col * width,
-      topRow: clarifaiface.top_row * height,
-      rightCol: width - (clarifaiface.right_col * width),
-      bottomRow: height - (clarifaiface.bottom_row * height)
-    }
+    const clarifaiface = data.outputs[0].data.regions;//[0].region_info.bounding_box;
+    // const image = document.getElementById('inputimage');
+    // const width = Number(image.width);
+    // const height = Number(image.height);
+    // console.log(width,height);
+    // return{
+    //   leftCol: clarifaiface.left_col * width,
+    //   topRow: clarifaiface.top_row * height,
+    //   rightCol: width - (clarifaiface.right_col * width),
+    //   bottomRow: height - (clarifaiface.bottom_row * height)
+    // }
+    return clarifaiface;
   }
 
   displayFaceBox = (box) => {
-    console.log(box);
+    // console.log(box[0]);
     this.setState({box});
   }
