+++ src/Components/FaceRecognition/FaceRecognition.js
@@ -1,12 +1,20 @@
 import React from 'react';
 import './FaceRecognition.css';
-
+import data from './calculate';
 const FaceRecognition = ({box,image}) =>{
     return(
         <div className="center ma">
             <div className="absolute mt2">
                 <img id="inputimage" alt="" src={image} width='500px' height='auto'/>
-                <div className="bounding-box" style={{top: box.topRow, right: box.rightCol, bottom: box.bottomRow, left: box.leftCol}}></div>
+                
+                <div>{ box.length ?
+                    box.map((element,i) => {
+                        let res = data(element.region_info);
+                        return <div className="bounding-box" key={i} style={{top: res.topRow, right: res.rightCol, bottom: res.bottomRow, left: res.leftCol}}></div>
+                    })
+                    : " "
+                }
+                </div>
             </div>
         </div>
     );
